[Memory64] Replace int (used as ptr) with long in tests
There are a lot of places where pointers get cast to int, and to prepare for further Memory64 changes, those are changed to long.
I chose long since it is most similar to the existing int, instead of e.g intptr_t, since long is more universal (there are many cases where the actual data stored is not a pointer).
Since under wasm32 int and long are equal, this should in theory be a NFC for wasm32, making it easier to review.
It also allows https://github.com/emscripten-core/emscripten/pull/12869 to be smaller

+++ tests/aniso.c
@@ -222,9 +222,9 @@ int main(int argc, char *argv[])
 
     glBegin( GL_TRIANGLE_STRIP );
     void (*f)(int, int) = glVertex2i;
-    if ((int)f % 16 == 4) f(5, 7);
+    if ((long)f % 16 == 4) f(5, 7);
     void (*g)(int, int) = glVertex3f;
-    if ((int)g % 16 == 4) g(5, 7);
-    return (int)f + (int)g;
+    if ((long)g % 16 == 4) g(5, 7);
+    return (int)((long)f + (long)g);
 }
