+++ tests/pthread/test_pthread_64bit_atomics.cpp
@@ -88,12 +88,12 @@ struct Test t[NUM_THREADS] = {};
 pthread_t thread[NUM_THREADS];
 
 void RunTest(int test)
-{	
+{
 	pthread_attr_t attr;
 	pthread_attr_init(&attr);
 	pthread_attr_setstacksize(&attr, 4*1024);
 
-	printf("Main thread has thread ID %d\n", (int)pthread_self());
+	printf("Main thread has thread ID %ld\n", pthread_self());
 	assert(pthread_self() != 0);
 
 	switch(test)
